blog/requirements.txt
appnope==0.1.0
backcall==0.1.0
cffi==1.11.5
click==6.7
decorator==4.3.0
Flask==1.0.2
Frozen-Flask==0.15
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pycparser==2.18
Pygments==2.2.0
PyYAML==3.12
ruamel.appconfig==0.5.3
ruamel.std.argparse==0.8.1
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
